Add script shaders
Groups can now contain custom shaders, which can be set to be used for a certain
mode.
The shaders are named <identifier>.<type>.<language>
<identifier> - the identifier which can be used in C4Script
<type> - the shader type. `vs` for a vertex shader, `fs` for a fragment shader.
<language> - the shader language. `glsl` for GLSL, `hlsl` for HLSL, `spirv` for
SPIR-V.

Keep in mind that script shaders are _optional_ -  the current graphics engine
does not have to support all of those languages or types. Thus, script shaders
are not synced.